+++ R/space_filling.R
@@ -224,7 +224,7 @@ make_sfd <- function(
 
     if (has_premade_design) {
       grid <- sfd::get_design(p, num_points = size, type = type)
-      vals <- purrr::map(params, \(.x) value_seq(.x, size))
+      vals <- purrr::map(params, \(.x) value_seq(.x, size, original = original))
       vals <- purrr::map(vals, \(.x) base_recycle(.x, size))
       grid <- sfd::update_values(grid, vals)
       names(grid) <- names(params)